A beginner-level challenge from Frontend Mentor,
The goal was to recreate a static component with accurate spacing, alignment, and responsiveness. I used semantic HTML, modern CSS (Flexbox and media queries), and a mobile-first approach to ensure the layout adapts across devices. This project helped refresh my frontend workflow and reinforce the basics of responsive design, accessibility, and component structure.